requirements.txt
# Microsoft Agent Framework
agent-framework

# Azure Identity
azure-identity

# Redis
redis

# Environment
python-dotenv
